+++ xcube_gen/api.py
@@ -58,13 +58,21 @@ class ApiResponse:
 
 
 class ApiError(BaseException):
-    def __init__(self, status_code: int, message: str):
+    def __init__(self, status_code: int,
+                 message: str,
+                 traceback: Optional[str] = None,
+                 output: Optional[str] = None):
         super().__init__(message)
         self.status_code = status_code
+        self._output = output
+        self._traceback = traceback
 
     @property
     def response(self):
-        return ApiResponse.error(error=self, status_code=self.status_code)
+        return ApiResponse.error(error=self,
+                                 status_code=self.status_code,
+                                 output=self._output,
+                                 traceback=self._traceback)
 
 
 def get_json_request_value(request: JsonObject,